[frontend] Delete kicked user from sidebar when user is kicked
